// front-web/src/pages/Records/types.ts
export type RecordResponse = {
    content: RecordItem[];
    totalPages: number;
}

export type RecordItem = {
    id: number;
    moment: string;
    name: string;
    age: number;
    gameTitle: Platform;
    gamePlatform: string;
    genreName: string;
}

export type Platform = 'XBOX' | 'PC' | 'PLAYSTATION';